Create VulnerabilityScanner.java
File 1 (Study_log4j.py) is a script that performs a detailed analysis of a log4j system. The script is responsible for identifying known vulnerabilities in the log4j system and sending them to File 2 for further analysis.
During the analysis, the script performs the following actions:

Sends a GET request to the log4j system to retrieve the headers.
Analyzes the headers for the presence of security headers and the encryption standard used.
Identifies known vulnerability patterns in the response.
Sends the vulnerabilities to File 2 for further analysis.
The script is designed to work in a flexible way and can be easily adjusted or extended to check for other types of vulnerabilities or to check for other types of headers. It is designed to be able to send the vulnerabilities to a second script or another endpoint